extract ga4gh query handler into its own class
